Feat: Listando y consultando por id un asesor

+++ database/migrations/2022_02_10_184228_create_asesor_table.php
@@ -18,7 +18,7 @@ class CreateAsesorTable extends Migration
             $table->string('nombre');
             $table->string('apellido');
             $table->string('tipo_documento');
-            $table->integer('numero_documento')->unique();
+            $table->bigInteger('numero_documento')->unique();
             $table->bigInteger('celular');
             $table->string('correo')->unique();
             $table->string('direccion');
